Fixed minor bug in mmu

+++ GB/Emulator/MMU.cs
@@ -108,7 +108,7 @@ namespace GB.Emulator
                 var a when a >= 0xff04 && a <= 0xff07 => clock.ReadByte(addr),
                 var a when a >= 0xff80 && a <= 0xfffe => HRAM[0x7f & addr],
                 var a when a >= 0xff10 && a <= 0xff26 => sound.ReadByte(a),
-                var a when a >= 0xff00 && a <= 0xff7f => a == 0xff00 ? 0xff : this.IOregisters[0xff & a],
+                var a when a >= 0xff00 && a <= 0xff7f => a == 0xff00 ? (byte)0xff : this.IOregisters[0xff & a],
                 var a when a >= 0xfea0 && a <= 0xfeff => 0xff
             };
             throw new NotImplementedException();
